pyproject.toml
[project]
name = "goob-brain"
version = "0.1.0"
description = "Discord-controlled Braccio arm with Claude vision."
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "anthropic>=0.40",
    "opencv-python>=4.8",
    "pyserial>=3.5",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]
# Voice frontend. Install with `pip install -e .[voice]`. Requires libportaudio
# system package on Linux (apt install libportaudio2 portaudio19-dev).
voice = [
    "sounddevice>=0.4",
    "numpy>=1.26",
    "openwakeword>=0.6",
    "webrtcvad-wheels>=2.0.14",
    "faster-whisper>=1.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
